requirements.txt
opencv-python>=4.8.0
google-cloud-vision>=3.4.4
pillow>=10.0.0
python-dotenv>=1.0.0